% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureDSR-package.R
\docType{package}
\name{AzureDSR-package}
\alias{AzureDSR}
\alias{AzureDSR-package}
\title{AzureDSR}
\description{
The AzureDSR functions boost efficiency of data science analytics with Azure resources.
}
\keyword{package}

