% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensionDSVM.R
\name{checkExtensionDSVM}
\alias{checkExtensionDSVM}
\title{Get information from a VM extension.}
\usage{
checkExtensionDSVM(context, resource.group, hostname)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}
}
\description{
Get information from a VM extension.
}

