% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenseCalculator.R
\name{costDSVM}
\alias{costDSVM}
\title{Calculate cost of using a specific DSVM instance of Azure for certain period.}
\usage{
costDSVM(context, hostname = "", timeStart, timeEnd, granularity = "Daily",
  currency = "USD", locale = "en-US", offerId, region = "US",
  verbose = FALSE)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{timeStart}{Start time.}

\item{timeEnd}{End time.}

\item{granularity}{Aggregation granularity. Can be either "Daily" or 
"Hourly".}

\item{currency}{Currency in which price rating is measured.}

\item{locale}{Locality information of subscription.}

\item{offerId}{Offer ID of the subscription. For more information see https://azure.microsoft.com/en-us/support/legal/offer-details/}

\item{region}{region information about the subscription.}
}
\value{
Total cost measured in the given currency of the specified DSVM 
instance in the period.
}
\description{
Note if difference between \code{timeStart} and \code{timeEnd} is less than 
the finest granularity, e.g., "Hourly" (we notice this is a usual case when 
one needs to be aware of the charges of a job that takes less than an hour),
the expense will be estimated based solely on computation hour. That is, 
the total expense is the multiplication of computation hour and pricing 
rate of the requested instance.
}
\seealso{
Other Cost.functions: \code{\link{dataConsumptionDSVM}}
}

