% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInterface.R
\name{createComputeInterface}
\alias{createComputeInterface}
\title{Create an compute interface object that handles interaction with remote instance.}
\usage{
createComputeInterface(remote, user, script, config)
}
\arguments{
\item{remote}{URL of remote instance.}

\item{user}{User name.}

\item{script}{R script with full path for execution at remote instance.}

\item{config}{Configuration for remote execution. Settings include computing context, data reference, etc.}
}
\value{
An S3 compute interface object.
}

