% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInterface.R
\name{createRInterface}
\alias{createRInterface}
\title{Create an R interface object that handles interaction with remote instance.}
\usage{
createRInterface(remote, user, script, config)
}
\arguments{
\item{remote}{URL of remote instance.}

\item{user}{User name.}

\item{script}{R script with full path for execution at remote instance.}

\item{config}{Configuration for remote execution. Settings include computing context, data reference, etc.}
}
\value{
An S3 R interface object.
}

