% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenseCalculator.R
\name{dataConsumption}
\alias{dataConsumption}
\title{Get data consumption of an Azure subscription for a time period. Aggregation method can be either daily based or hourly based.}
\usage{
dataConsumption(context, instance, timeStart, timeEnd, granularity = "Hourly")
}
\arguments{
\item{context}{AzureSMR context object.}

\item{instance}{Instance of Azure instance name that one would like to check expense. For example, it can be the name of a virtual machine.}

\item{timeStart}{Start time.}

\item{timeEnd}{End time.}

\item{granularity}{Aggregation granularity. Can be either "Daily" or "Hourly".}
}
\note{
Formats of start time point and end time point follow ISO 8601 standard. Say if one would like to calculate data consumption between Feb 21, 2017 to Feb 25, 2017, the inputs should be "2017-02-21 00:00:00" and "2017-02-25 00:00:00", for start time point and end time point, respectively. For hourly based calculation, note there should be no minute and second included.
}

