% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployDSVMCluster.R
\name{deployDSVMCluster}
\alias{deployDSVMCluster}
\title{This function creates a cluster of Data Science Virtual Machines and enable the DSVMs to communicate across the cluster via public key based credentials for high performance computation. All DSVMs in the cluster are based on Linux OS and use public key cryptogrphy for log in.}
\usage{
deployDSVMCluster(context, resource.group, location, count = 1, name,
  username, size = "Standard_D1_v2", pubkey = "", dns = name,
  cluster = FALSE)
}
\arguments{
\item{context}{AzureSMR active context.}

\item{resource.group}{The Azure resource group where the DSVM is allocated.}

\item{location}{Location of the data centre to host the DSVM.}

\item{count}{Number of DSVM instances to be created. Note deploying multiple DSVMs may consume some time.}

\item{name}{Names of the DSVMs. Lowercase characters or numbers only. Special characters are not permitted.}

\item{username}{User name of the DSVM. It should be different from name of the DSVM.}

\item{size}{Size of the DSVM cluster is identical.}

\item{pubkey}{Public key for the DSVM. Only applicable for
public-key based authentication of Linux based DSVM.}

\item{dns}{DNS label for the VM address. The URL for accessing the deployed DSVM will be "<dns_label>.<location>.cloudapp.azure.com}

\item{cluster}{A logical value of TRUE or FALSE to indicate whether the deployed DSVMs form a cluster. If not, the deployment will assign the vectors of name, username, and public key as given in the input arguments to the DSVMs - this is usually used for creating multiple DSVMs for a group of data scientists. If it is TRUE, the deployment will use the first element (if it consists more than one elements) of the given DSVM names as base, and append serial number to the base to form a DSVM full name, and then use the SAME username and public key across the cluster - this can be used for creating a HPC engine on top of the deployed DSVMs in which parallel computing context which is availed in Microsoft R Server ScaleR package can be applied for embarassing parallelization.}
}

