% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInterface.R
\name{dumpInterface}
\alias{dumpInterface}
\title{Dump the interface configuration.}
\usage{
dumpInterface(object)
}
\arguments{
\item{object}{The compute interface object.}
}
\value{
No return. Print compute interface object information.
}

