% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpInterface.R
\name{dumpObject}
\alias{dumpObject}
\title{Dump out the object configuration.}
\usage{
dumpObject(object)
}
\arguments{
\item{object}{The R interface object.}
}
\value{
No return. Print R interface object information.
}

