% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeScript.R
\name{executeScript}
\alias{executeScript}
\title{Remote execution of R script in an R interface new_interface.}
\usage{
executeScript(context, resource.group, hostname, remote, username, script,
  master = "", slaves = "", compute.context = ifelse(master == "" &&
  slaves == "", "localParallel", "clusterParallel"))
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{remote}{IP address or FQDN for a computation engine. For
DSVM, it is either the fully qualified domain name (usually in 
the format of
<hostname>.<location>.cloudapp.azure.com) or its public IP
address. Note if more than one machines are used for execution,
the remote is used as master node by default.}

\item{username}{User name of the DSVM. It should be different from
`name`.}

\item{script}{R script to be executed on remote resource(s).}

\item{master}{IP address or URL of a DSVM which will be used as the
master. By default is remote.}

\item{slaves}{IP addresses or URLs of slave DSVMs.}

\item{compute.context}{Computation context of Microsoft R Server
under which the mechanisms of parallelization (e.g., local
parallel, cluster based parallel, etc.) is
specified. Accepted computing context include "localSequential", 
"localParallel", and
"clusterParallel".}
}
\value{
Status of scription execution.
}
\description{
Remote execution of R script in an R interface new_interface.
}
\details{
For a "localSequential" compute context, there is no parallelism and 
the analytics run in a sequential manner. In the "localParallel" 
compute context, the analytics will be run across available cores 
of a computing node. The number of available cores can be checked with
 Microsoft RevoScaleR function \code{rxGetOption}. The "clusterParallel" 
 compute context enables parallel computation across coputing nodes of
  a cluster.
}
\references{
Details about distributed computing with Microsoft 
RevoScaleR package can be found at 
https://msdn.microsoft.com/en-us/microsoft-r/scaler-distributed-computing.
}

