% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existsRG.R
\name{existsRG}
\alias{existsRG}
\title{Check if a resource group exists.}
\usage{
existsRG(context, resource.group, location, verbose = TRUE)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{location}{Location of the data centre to host the DSVM.}
}

