% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenseCalculator.R
\name{expenseCalculator}
\alias{expenseCalculator}
\title{Calculate cost of using a specific instance of Azure for certain period.}
\usage{
expenseCalculator(context, instance, timeStart, timeEnd, granularity, currency,
  locale, offerId, region)
}
\arguments{
\item{context}{AzureSMR context.}

\item{instance}{Instance of Azure instance that one would like to check expense. No matter whether resource group is given or not, if a instance of instance is given, data consumption of that instance is returned.}

\item{timeStart}{Start time.}

\item{timeEnd}{End time.}

\item{granularity}{Aggregation granularity. Can be either "Daily" or "Hourly".}

\item{currency}{Currency in which price rating is measured.}

\item{locale}{Locality information of subscription.}

\item{offerId}{Offer ID of the subscription. Detailed information can be found at https://azure.microsoft.com/en-us/support/legal/offer-details/}

\item{region}{region information about the subscription.}
}
\value{
Total cost measured in the given currency of the specified Azure instance in the period.
}

