% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeScript.R
\name{fileTransfer}
\alias{fileTransfer}
\title{Upload or download files.}
\usage{
fileTransfer(from = ".", to = ".", user, file)
}
\arguments{
\item{from}{Source location (path) of file.}

\item{to}{Target location (path) of file.}

\item{file}{File name - a character string.}
}
\note{
File transfer is implemented by `scp` with public key based authentication.
}

