% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVMSizes.R
\name{getVMSizes}
\alias{getVMSizes}
\title{Get available sizes for data science virtual machines.}
\usage{
getVMSizes(context, location)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{location}{Location of the data centre to host the DSVM.}
}
\value{
A data frame that contains basic information such as number of cores, disk size, RAM size, etc.,
about the available DSVM sizes.
}
\description{
Get available sizes for data science virtual machines.
}

