% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyDistribution.R
\name{keyDistribution}
\alias{keyDistribution}
\title{Distribute public keys across nodes.}
\usage{
keyDistribution(location, hostnames, usernames, pubkeys, count,
  dns.labels = hostnames)
}
\arguments{
\item{location}{Location of the DSVMs.}

\item{hostnames}{Host names of the DSVMs.}

\item{usernames}{User name used for the DSVMs. In the case of DSVM cluster, the user name is unique.}

\item{pubkeys}{Public keys used for the DSVMs.}

\item{dns.labels}{DNS labels for the DSVMs.}
}
\value{
No returned value.
}
\description{
The function distributes public keys across nodes so as to allow secure communication between each of the nodes. This is to enable doParallel backend to create socket-based cluster based on the nodes for parallel computation.
}
\note{
Note this function merely applies for Linux DSVM. In addition, it relies on SSH/SCP for file transfer so OpenSSH or other SSH tools are required if they do not exist by default in the system.
}

