% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneBoxConfiguration.R
\name{mrsOneBoxConfiguration}
\alias{mrsOneBoxConfiguration}
\title{One box configuration for VM with Microsoft R Server (>= 9.1)}
\usage{
mrsOneBoxConfiguration(context, resource.group, hostname, location, username,
  password)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{location}{Location of the data centre to host the DSVM.}

\item{username}{User name of the DSVM. It should be different from
`name`.}

\item{password}{The password used for the one-box configuration. Default 
username for the configuration is 'admin'.}
}
\description{
The function will perform one-box configuration for a MRS VM.
A successful configuration will allow one to operationalize MRS analytics
on a VM with mrsdeploy functions.
}
\references{
More can be found at 
https://msdn.microsoft.com/en-us/microsoft-r/operationalize/configuration-initial
}

