% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newScript.R
\name{newScript}
\alias{newScript}
\title{Generate a new worker script which is run on the remote instance with specifications in R interface object configuration.}
\usage{
newScript(path = ".", title = paste0("worker_new_", Sys.time(), ".R"))
}
\arguments{
\item{path}{Path to the script.}

\item{title}{Title of the script.}
}

