% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenseCalculator.R
\name{pricingRates}
\alias{pricingRates}
\title{Get pricing details of resources under a subscription.}
\usage{
pricingRates(context, currency, locale, offerId, region)
}
\arguments{
\item{context}{- Azure Context Object.}

\item{currency}{Currency in which price rating is measured.}

\item{locale}{Locality information of subscription.}

\item{offerId}{Offer ID of the subscription. Detailed information can be found at https://azure.microsoft.com/en-us/support/legal/offer-details/}

\item{region}{region information about the subscription.}
}

