% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInterface.R
\name{setConfig}
\alias{setConfig}
\title{Set configuration for the compute interface object.}
\usage{
setConfig(object, machine_list, dns_list, machine_user, master, slaves, data,
  context)
}
\arguments{
\item{object}{An S3 compute interface object.}

\item{machine_list}{List of remote instances that execute R scripts.}

\item{dns_list}{DNS of the remote instances.}

\item{machine_user}{User name of the remote instances.}

\item{master}{Master node of the machine.}

\item{slaves}{Slave nodes of the machine.}

\item{data}{Reference to data used in the analytics.}

\item{context}{Computing context available in Microsoft R Server for running the analytics.}
}

