% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInterface.R
\name{setRInterface}
\alias{setRInterface}
\title{Set values of the R interface object.}
\usage{
setRInterface(object, remote, user, script, config)
}
\arguments{
\item{object}{An S3 R interface object.}

\item{remote}{URL of remote instance.}

\item{user}{User name.}

\item{script}{R script with full path for execution at remote instance.}

\item{config}{Configuration for remote execution. Settings include computing context, data reference, etc.}
}
\value{
The updated R interface object.
}

