% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInterface.R
\name{updateScript}
\alias{updateScript}
\title{Update a worker script with compute interface object configuration.}
\usage{
updateScript(object)
}
\arguments{
\item{object}{compute interface object.}
}

