% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensionDSVM.R
\name{addExtensionDSVM}
\alias{addExtensionDSVM}
\title{Add DSVM extension.}
\usage{
addExtensionDSVM(context, location, resource.group, hostname, fileurl, command)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{location}{Location of the data centre to host the DSVM.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{fileurl}{The URL of script to execute.}

\item{command}{Command to run.}
}
\description{
Add DSVM extension.
}

