% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumptionCalculator.R
\name{dataConsumption}
\alias{dataConsumption}
\title{Get data consumption of an Azure subscription for a time period between two time points. The granularity of the time can be either daily based or hourly based.}
\usage{
dataConsumption(context, resourceGroup, instance, timeStart, timeEnd,
  granularity)
}
\arguments{
\item{context}{AzureSMR context object.}

\item{resourceGroup}{Resource group that one would like to check. If a resource group is given, instance of instance can be omitted. In this case, all data consumption records under the resource group will be returned.}

\item{instance}{Instance of Azure instance that one would like to check expense. No matter whether resource group is given or not, if a instance of instance is given, data consumption of that instance is returned.}

\item{timeStart}{Start time.}

\item{timeEnd}{End time.}

\item{granularity}{Aggregation granularity. Can be either "Daily" or "Hourly".}
}
\note{
Formats of start time point and end time point follow ISO 8601 standard. Say if one would like to calculate data consumption between Feb 21, 2017 to Feb 25, 2017, the inputs should be "2017-02-21 00:00:00" and "2017-02-25 00:00:00", for start time point and end time point, respectively. For hourly based calculation, note there should be no minute and second included.
}

