% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployDSVM.R
\name{deployDSVM}
\alias{deployDSVM}
\title{Deploy a new Data Science Virtual Machine (DSVM).}
\usage{
deployDSVM(context, resource.group, location, vmname, vmusername,
  vmsize = "Standard_D3_v2", vmos, vmauthen = "", vmpubkey = "",
  vmpassword = "", vmdns = paste0(vmname, "dns"), mode = "Sync")
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{location}{Location of the data centre to host the DSVM.}

\item{vmname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{vmusername}{User name of the DSVM. It should be different
from `vmname`.}

\item{vmsize}{Size of the DSVM. The default is
"Standard_D1_v2". All available sizes can be obtained by function
`getVMSizes`.}

\item{vmos}{Operating system of DSVM. Permitted values are "Linux" and "Windows" for Linux based and Windows based operating systems, respectively.}

\item{vmauthen}{Either "Key" or "Password", meaning public-key based or
password based authentication, respectively. Note Windows DSVM by default uses password based authentication and this argument can be left unset.}

\item{vmpubkey}{Public key for the DSVM. Only applicable for
public-key based authentication of Linux based DSVM.}

\item{vmpassword}{Pass word for the DSVM.}

\item{vmdns}{DNS label for the VM address. The URL for accessing the deployed DSVM will be "<dns_label>.<location>.cloudapp.azure.com}

\item{mode}{Mode of virtual machine deployment. Default is "Sync".}
}

