% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployDSVM-MININT-AL1V820.R, R/deployDSVM.R
\name{deployDSVM}
\alias{deployDSVM}
\title{Deploy a new Data Science Virtual Machine (DSVM).}
\usage{
deployDSVM(context, resource.group, location, name, username,
  size = "Standard_D1_v2", os, authen = "", pubkey = "", password = "",
  dns = name, mode = "Sync")

deployDSVM(context, resource.group, location, name, username,
  size = "Standard_D1_v2", os, authen = "", pubkey = "", password = "",
  dns = name, mode = "Sync")
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{location}{Location of the data centre to host the DSVM.}

\item{name}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{username}{User name of the DSVM. It should be different from
`name`.}

\item{size}{Size of the DSVM. The default is "Standard_D1_v2". All
available sizes can be obtained by function `getVMSizes`.}

\item{os}{Operating system of DSVM. Permitted values are "Linux"
and "Windows" for Linux based and Windows based operating
systems, respectively.}

\item{authen}{Either "Key" or "Password", meaning public-key based
or password based authentication, respectively. Note Windows DSVM
by default uses password based authentication and this argument
can be left unset.}

\item{pubkey}{Public key for the DSVM. Only applicable for
public-key based authentication of Linux based DSVM.}

\item{password}{Pass word for the DSVM.}

\item{dns}{DNS label for the VM address. The URL for accessing the
deployed DSVM will be "<dns_label>.<location>.cloudapp.azure.com}

\item{mode}{Mode of virtual machine deployment. Default is "Sync".}

\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{location}{Location of the data centre to host the DSVM.}

\item{name}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{username}{User name of the DSVM. It should be different from
`name`.}

\item{size}{Size of the DSVM. The default is "Standard_D1_v2". All
available sizes can be obtained by function `getVMSizes`.}

\item{os}{Operating system of DSVM. Permitted values are "Linux"
and "Windows" for Linux based and Windows based operating
systems, respectively.}

\item{authen}{Either "Key" or "Password", meaning public-key based
or password based authentication, respectively. Note Windows DSVM
by default uses password based authentication and this argument
can be left unset.}

\item{pubkey}{Public key for the DSVM. Only applicable for
public-key based authentication of Linux based DSVM.}

\item{password}{Pass word for the DSVM.}

\item{dns}{DNS label for the VM address. The URL for accessing the
deployed DSVM will be "<dns_label>.<location>.cloudapp.azure.com}

\item{mode}{Mode of virtual machine deployment. Default is "Sync".}
}

