% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployDSVM.R
\name{deployDSVM}
\alias{deployDSVM}
\title{Deploy a new Data Science Virtual Machine (DSVM).}
\usage{
deployDSVM(context, resource.group, location, hostname, username,
  size = "Standard_D1_v2", os = "Linux", authen = ifelse(os == "Linux",
  "Key", "Password"), pubkey = "", password = "", dns.label = hostname,
  mode = "Sync")
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{location}{Location of the data centre to host the DSVM.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{username}{User name of the DSVM. It should be different from
`name`.}

\item{size}{Size of the DSVM. The default is "Standard_D1_v2". All
available sizes can be obtained by function `getVMSizes`.}

\item{os}{Operating system of DSVM. Permitted values are "Linux"
and "Windows". The default is to deploy a Linux Data Science
Virtual Machine.}

\item{authen}{Either "Key" for public-key based authentication
(with Linux) or "Password" for a password based authentication
(Linux or Windows). Default is to use public key authentication
for Linux and password based authentication for Windows.}

\item{pubkey}{Public key for the DSVM. Only applicable for
public-key based authentication of Linux based DSVM.}

\item{password}{Pass word for the DSVM.}

\item{dns.label}{DNS label for the VM address. The URL for
accessing the deployed DSVM will be
"<dns_label>.<location>.cloudapp.azure.com}

\item{mode}{Mode of virtual machine deployment. Default is "Sync".}
}
\details{
Usually an error message will be returned if the deployment is unsuccessful. If the deployment fails without explicit error message returned visit the Azure portal
https://ms.portal.azure.com and browse to the resource group and
click on the failed deployment link to view the failure
message. Typical errors include DnsRecordInUse or
StorageAccountAlreadyTaken. If so then choose a different hostname.
}

