% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployDSVMCluster.R
\name{deployDSVMCluster}
\alias{deployDSVMCluster}
\title{Deploy a cluster of Linux Data Science Virtual Machines on Azure.}
\usage{
deployDSVMCluster(context, resource.group, location, hostnames, usernames,
  pubkeys, count, size = "Standard_D1_v2", dns.labels = hostnames)
}
\arguments{
\item{context}{AzureSMR active context.}

\item{resource.group}{The Azure resource group where the DSVM is
allocated.}

\item{location}{Location of the data centre to host the DSVM.}

\item{hostnames}{Hostnames for the DSVMs. Lowercase characters or
numbers only. If a single hostname is supplied and count > 1 then
the hostname will be used as a prefix for a sequential count of
hostnames.}

\item{usernames}{Usernames for the admin account created on the
DSVM. If a single username is supplied then that username is used
as the admin user on each host. Otherwise a username is provided
for each of the DSVMs.}

\item{pubkeys}{Public keys for the DSVM. This is only applicable
for public-key based authentication of Linux based DSVM. One or a
vector of public keys can be provided, depending on the count or
the number of hostnames or usernames.}

\item{count}{If provided this is the number of DSVM instances to be
created. If not provided the number of DSVMs created will be
either the number of names provided or the number of usernames
provided.}

\item{size}{The size of the DSVMs. Each DSVM is the same size.}

\item{dns.labels}{DNS labels for the VM. By default this is the
hostnames but is not required to be. The fully qualified domain
name for accessing the deployed DSVM will then be
"<dns.label_label>.<location>.cloudapp.azure.com".}
}
\description{
Creates a cluster of Data Science Virtual Machines and enables the
DSVMs to communicate across the cluster via public key based
credentials for high performance computation. All DSVMs in the
cluster are based on Linux OS and use public key cryptogrphy for
log in.
}
\details{
We identify two specific use cases but recognise there are many
that are supported by this function.

A cluster is intended as a High Performance Compute engine across
the deployed DSVMs supporting a parallel computing context as is
available with Microsoft R Server ScaleR package. A cluster is a
deployment of multiple identitical DSVMs. A single admin username
and public key will be used across the cluster. The individual
machine names will be constructed from the provided name with
sequential numbers. The data scientist will typically connect to
the cluster from their local desktop/laptop running R locally with
remote execution for computation. A cluster is typcially created by
the data scientist when needed and the resource group deleted on
completion of the activity.

A collection is a deployment with different usernames and public
keys for each of the DSVMS. A vector of usernames must be
provided. A colleciton is often used in creating multiple DSVMs for
a group of data scientists or for training. A colleciton is often
longer lasting than a cluster.
}

