% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployDSVMCluster.R
\name{deployDSVMCluster}
\alias{deployDSVMCluster}
\title{Deploy a cluster of Linux Data Science Virtual Machines on Azure.}
\usage{
deployDSVMCluster(context, resource.group, location, hostname, username,
  authen = "Key", pubkey = "", password = "", os = "Ubuntu",
  size = "Standard_D1_v2", dns.label = hostname, count)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{resource.group}{The Azure resource group where the DSVM is
created.}

\item{location}{Location of the data centre to host the DSVM.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{username}{User name of the DSVM. It should be different from
`name`.}

\item{authen}{Either "Key" for public-key based authentication
(with Linux) or "Password" for a password based authentication
(Linux or Windows). Default is to use public key authentication
for Linux and password based authentication for Windows.}

\item{pubkey}{Public key for the DSVM. Only applicable for
public-key based authentication of Linux based DSVM.}

\item{password}{Pass word for the DSVM.}

\item{os}{Operating system of DSVM. Permitted values are "Ubuntu",
"CentOS", "Windows", "DeepLearning", and "MRS". The default is to deploy
a Ubuntu Linux Data Science Virtual Machine. NOTE previously Windows
based DSVM for deep learning toolkit is using a separated package but
now it is merged into one.}

\item{size}{Size of the DSVM. The default is "Standard_D1_v2". All
available sizes can be obtained by function `getVMSizes`.}

\item{dns.label}{DNS label for the VM address. The URL for
accessing the deployed DSVM will be
"<dns_label>.<location>.cloudapp.azure.com}

\item{count}{If provided this is the number of DSVM instances to be
created. If not provided the number of DSVMs created will be
either the number of names provided or the number of username
provided.}
}
\description{
Creates a cluster of Data Science Virtual Machines and enables the
DSVMs to communicate acros the cluster via public key based
credentials for high performance computation. All DSVMs in the
cluster are based on Linux OS and use public key cryptogrphy for
log in.
}
\details{
Note clustering of DSVMs for high performance computing will be 
deprecated. Users are encouraged to use azureDoParallel package 
for the same purpose with Azure Batch Services. 

The current function supports for deployment of identical DSVMs 
(i.e., size, OS, etc.) or heterogeneous DSVMs with differerent 
specifications.
}
\examples{
\dontrun{
# The following deploys a cluster of 5 Linux DSVMs.

deployDSVMCluster(context, resource.group="<resource_group>", 
location="<location>", hostname="<machine_name>", username="<user_name>", 
os="Windows", pubkey="<a_valid_public_key_string_in_SSH_format>", count=5)

# The following deploys a collection of 3 Linux DSVMs with different 
names and public keys.

deployDSVMCluster(context, resource.group="<resource_group>", 
location="<location>", hostname=c("<machine_name_1>", "<machine_name_2>", 
"<machine_name_3>", username=c("<user_name_1>", "<user_name_2>", 
"<user_name_3>"), os="Windows", 
pubkey="<a_valid_public_key_string_in_SSH_format>"}
}

