% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeScript.R
\name{executeScript}
\alias{executeScript}
\title{Remote execution of R script in an R interface object.}
\usage{
executeScript(object, inputs = NULL, outputs = NULL,
  checkLibraries = FALSE, displayPlots = FALSE, writePlots = FALSE)
}
\arguments{
\item{object}{R interface object.}

\item{inputs}{JSON encoded string of R objects that are loaded into the Remote R session's workspace prior to execution. Only R objects of type: primitives, vectors and dataframes are supported via this parameter. Alternatively the putLocalObject can be used, prior to a call to this function, to move any R object from the local workspace into the remote R session.}

\item{outputs}{Character vector of the names of the objects to retreive. Only primitives, vectors and dataframes can be retrieved using this function. Use getRemoteObject to get any type of R object from the remote session.}

\item{checkLibraries}{if `TRUE`, check whether libraries used in the R script installed on the remote machine.}

\item{displayPlots}{If TRUE, plots generated during execution are displayed in the local plot window. **NOTE** This capability requires that the 'png' package is installed on the local machine.}

\item{writePlots}{If TRUE, plots generated during execution are copied to the working directory of the local session.}
}
\value{
Status of scription execution.
}

