% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeScript.R
\name{executeScript}
\alias{executeScript}
\title{Remote execution of R script in an R interface new_interface.}
\usage{
executeScript(context, resourceGroup, machines, remote, user, script, master,
  slaves, computeContext)
}
\arguments{
\item{context}{AzureSMR context.}

\item{resourceGroup}{Resource group of Azure resources for computation.}

\item{machines}{Remote DSVMs that will be used for computation.}

\item{remote}{IP address or URL for a computation engine. For DSVM, it is either DNS (usually in the format of <dsvm name>.<location>.cloudapp.azure.com) or its public IP address. Note if more than one machines are used for execution, the remote is used as master node by default.}

\item{user}{Username for logging into the remote resource.}

\item{script}{R script to be executed on remote resource(s).}

\item{master}{IP address or URL of a DSVM which will be used as the master. By default is remote.}

\item{slaves}{IP addresses or URLs of slave DSVMs.}

\item{computeContext}{Computation context of Microsoft R Server under which the mechanisms of parallelization (e.g., local parallel, cluster based parallel, or Spark) is specified. Accepted computing context include "localParallel", "clusterParallel", "Hadoop", and "Spark".}
}
\value{
Status of scription execution.
}

