% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeScript.R
\name{executeScript}
\alias{executeScript}
\title{Remote execution of R script in an R interface new_interface.}
\usage{
executeScript(context, resourceGroup, machines, remote, user, script, master,
  slaves, computeContext)
}
\arguments{
\item{context}{AzureSMR context.}

\item{resourceGroup}{Resource group of Azure resources for computation.}

\item{machines}{Remote DSVMs that will be used for computation.}

\item{remote}{IP address or FQDN for a computation engine. For
DSVM, it is either the fully qualified domain name (usually in the format of
<hostname>.<location>.cloudapp.azure.com) or its public IP
address. Note if more than one machines are used for execution,
the remote is used as master node by default.}

\item{user}{Username for logging into the remote resource.}

\item{script}{R script to be executed on remote resource(s).}

\item{master}{IP address or URL of a DSVM which will be used as the
master. By default is remote.}

\item{slaves}{IP addresses or URLs of slave DSVMs.}

\item{computeContext}{Computation context of Microsoft R Server
under which the mechanisms of parallelization (e.g., local
parallel, cluster based parallel, etc.) is
specified. Accepted computing context include "localSequential", "localParallel", and
"clusterParallel".}
}
\value{
Status of scription execution.
}
\description{
Remote execution of R script in an R interface new_interface.
}
\details{
For a "localSequential" compute context, there is no parallelism and the analytics run in a sequential manner. In the "localParallel" compute context, the analytics will be run across available cores of a computing node. The number of available cores can be checked with Microsoft RevoScaleR function \code{rxGetOption}. The "clusterParallel" compute context enables parallel computation across coputing nodes of a cluster.
}
\note{
Remote code/script execution by using Microsoft R Server 9.0, which brings more convenience and is portable to both Linux and Windows system, will be supported soon.
}
\references{
Details about distributed computing with Microsoft RevoScaleR package can be found at https://msdn.microsoft.com/en-us/microsoft-r/scaler-distributed-computing.
}

