% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeScript.R
\name{fileTransfer}
\alias{fileTransfer}
\title{Upload or download files.}
\usage{
fileTransfer(from = ".", to = ".", user, file)
}
\arguments{
\item{from}{Source location (local path or remote FQDN and path) of file.}

\item{to}{Target location (local path or remote FQDN and path) of file.}

\item{file}{File name - a character string.}
}
\note{
File transfer is implemented by `scp` with public key based 
authentication so it is limited only to Linux based DSVM at the moment.
}
\examples{
\dontrun{
# copy a file named "script.R" from local current working directory to 
remote (e.g., IP address is 192.168.19.1) home directory.
fileTransfer(from=".", to="192.168.19.1:~", user="admin", file="script.R")}
}

