% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getVMSizes}
\alias{getVMSizes}
\title{Get available sizes for data science virtual machines.}
\usage{
getVMSizes(context, location)
}
\arguments{
\item{context}{An AzureSMR context.}

\item{location}{Location of the Azure resources.}
}
\value{
A data frame that contains basic information about the available DSVM sizes.
}

