% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyDistribution.R
\name{keyDistribution}
\alias{keyDistribution}
\title{Distribute public keys across nodes.}
\usage{
keyDistribution(location, hostname, username, count, dns.labels = hostname)
}
\arguments{
\item{location}{Location of the data centre to host the DSVM.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{username}{User name of the DSVM. It should be different from
`name`.}

\item{count}{If provided this is the number of DSVM instances to be
created. If not provided the number of DSVMs created will be
either the number of names provided or the number of username
provided.}
}
\value{
No returned value.
}
\description{
The function distributes public keys across nodes so as to allow secure communication between each of the nodes. This is to enable doParallel backend to create socket-based cluster based on the nodes for parallel computation.
}
\note{
Note this function merely applies for Linux DSVM. In addition, it relies on SSH/SCP for file transfer so OpenSSH or other SSH tools are required if they do not exist by default in the system.
}

