% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneBoxConfiguration.R
\name{mrsOneBoxConfiguration}
\alias{mrsOneBoxConfiguration}
\title{One box configuration for VM with Microsoft R Server (>= 9.1)}
\usage{
mrsOneBoxConfiguration(context, hostname, password)
}
\arguments{
\item{context}{Authentication context of AzureSMR encapsulating the
TID, CID, and key obtained from Azure Actrive Directory.}

\item{hostname}{Name of the DSVM.  Lowercase characters or numbers
only. Special characters are not permitted.}

\item{password}{Pass word for the DSVM.}
}
\description{
One box configuration for VM with Microsoft R Server (>= 9.1)
}

