% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operateDSVM.R
\name{operateDSVM}
\alias{operateDSVM}
\title{Operations on a data science virtual machine. Available operations are "Check", "Start", "Stop", and "Delete".}
\usage{
operateDSVM(context, resource.group, names, operation = "Check")
}
\arguments{
\item{context}{AzureSMR context.}

\item{resource.group}{Resource group.}

\item{operation}{Operations on the DSVM. Available operations are "Check", "Start", "Stop", "Delete", which check the status of, start running, stop running, and delete a DSVM, respectively.}

\item{name}{Name(s) of the DSVM(s).}
}

